Initial commit for framework.
Turns run through scenes, per player. Scenes have three phases: start,
action, and end. Each phase is asynchronous, so that the game can pause
for the player in any of the phases.

Scenes are probably locked to specific locations in most cases, although
it seems okay to let them not be always locked to them.